feat: add `popularSearches` and `querySuggestions` endpoint adapters and rename `topClicked` to `recommendations`
EX-6078
