+++ packages/x-adapter-platform/src/__tests__/__fixtures__/recommendations.response.ts
@@ -1,4 +1,4 @@
-export const platformTopClickedResponse = {
+export const platformRecommendationsResponse = {
   topclicked: {
     content: [
       {
