--- packages/x-adapter-platform/src/endpoint-adapters/empathize.endpoint-adapter.ts
@@ -1,13 +1,0 @@
-import { endpointAdapterFactory, EndpointAdapterOptions } from '@empathyco/x-adapter-next';
-import { baseRequestMapper } from '../mappers/request/base-request.mapper';
-import { empathizeResponseMapper } from '../mappers/response/empathize-response.mapper';
-import { BaseRequest } from '../types/request.types';
-import { EmpathizeResponse } from '../types/response.types';
-
-const endpointAdapterOptions: EndpointAdapterOptions<BaseRequest, EmpathizeResponse> = {
-  endpoint: 'https://api.{env(.)}empathy.co/search/v1/query/{instance}/empathize',
-  responseMapper: empathizeResponseMapper,
-  requestMapper: baseRequestMapper
-};
-
-export const empathizeEndpointAdapter = endpointAdapterFactory(endpointAdapterOptions);
