+++ packages/x-adapter-platform/src/endpoint-adapters/index.ts
@@ -1,7 +1,8 @@
-export * from './empathize.endpoint-adapter';
+export * from './identifier-results.endpoint-adapter';
 export * from './next-queries.endpoint-adapter';
+export * from './popular-searches.endpoint-adapter';
+export * from './query-suggestions.endpoint-adapter';
+export * from './recommendations.endpoint-adapter';
 export * from './related-tags.endpoint-adapter';
 export * from './search.endpoint-adapter';
-export * from './sku-search.endpoint-adapter';
 export * from './tagging.endpoint-adapter';
-export * from './top-clicked.endpoint-adapter';
