+++ packages/x-adapter-platform/src/endpoint-adapters/next-queries.endpoint-adapter.ts
@@ -1,16 +1,13 @@
-import { endpointAdapterFactory, EndpointAdapterOptions } from '@empathyco/x-adapter-next';
-import { BaseRequest } from '../types/request.types';
-import { NextQueriesResponse } from '../types/response.types';
+import { endpointAdapterFactory } from '@empathyco/x-adapter-next';
+import { NextQueriesRequest, NextQueriesResponse } from '@empathyco/x-types';
+import { nextQueriesRequestMapper } from '../mappers/request/next-queries-request.mapper';
 import { nextQueriesResponseMapper } from '../mappers/response/next-queries-response.mapper';
-import { baseRequestMapper } from '../mappers/request/base-request.mapper';
 
-export const nextQueriesEndpointAdapterOptions: EndpointAdapterOptions<
-  BaseRequest,
+export const nextQueriesEndpointAdapter = endpointAdapterFactory<
+  NextQueriesRequest,
   NextQueriesResponse
-> = {
+>({
   endpoint: 'https://api.{env(.)}empathy.co/nextqueries/{instance}',
-  responseMapper: nextQueriesResponseMapper,
-  requestMapper: baseRequestMapper
-};
-
-export const nextQueriesEndpointAdapter = endpointAdapterFactory(nextQueriesEndpointAdapterOptions);
+  requestMapper: nextQueriesRequestMapper,
+  responseMapper: nextQueriesResponseMapper
+});
