+++ packages/x-adapter-platform/src/endpoint-adapters/related-tags.endpoint-adapter.ts
@@ -1,16 +1,13 @@
-import { endpointAdapterFactory, EndpointAdapterOptions } from '@empathyco/x-adapter-next';
-import { BaseRequest } from '../types/request.types';
-import { RelatedTagsResponse } from '../types/response.types';
+import { endpointAdapterFactory } from '@empathyco/x-adapter-next';
+import { RelatedTagsRequest, RelatedTagsResponse } from '@empathyco/x-types';
+import { relatedTagsRequestMapper } from '../mappers/request/related-tags-request.mapper';
 import { relatedTagsResponseMapper } from '../mappers/response/related-tags-response.mapper';
-import { baseRequestMapper } from '../mappers/request/base-request.mapper';
 
-export const relatedTagsEndpointAdapterOptions: EndpointAdapterOptions<
-  BaseRequest,
+export const relatedTagsEndpointAdapter = endpointAdapterFactory<
+  RelatedTagsRequest,
   RelatedTagsResponse
-> = {
+>({
   endpoint: 'https://api.{env(.)}empathy.co/relatedtags/{instance}',
-  responseMapper: relatedTagsResponseMapper,
-  requestMapper: baseRequestMapper
-};
-
-export const relatedTagsEndpointAdapter = endpointAdapterFactory(relatedTagsEndpointAdapterOptions);
+  requestMapper: relatedTagsRequestMapper,
+  responseMapper: relatedTagsResponseMapper
+});
