+++ packages/x-adapter-platform/src/endpoint-adapters/tagging.endpoint-adapter.ts
@@ -1,16 +1,9 @@
-import {
-  beaconHttpClient,
-  endpointAdapterFactory,
-  EndpointAdapterOptions,
-  buildUrl
-} from '@empathyco/x-adapter-next';
-import { TaggingRequest } from '../types/request.types';
+import { beaconHttpClient, endpointAdapterFactory, buildUrl } from '@empathyco/x-adapter-next';
+import { TaggingRequest } from '@empathyco/x-types';
 import { taggingRequestMapper } from '../mappers/request/tagging-request.mapper';
 
-export const taggingEndpointAdapterOptions: EndpointAdapterOptions<TaggingRequest, void> = {
+export const taggingEndpointAdapter = endpointAdapterFactory<TaggingRequest, void>({
   endpoint: ({ url, params }) => buildUrl(url, params),
   httpClient: beaconHttpClient,
   requestMapper: taggingRequestMapper
-};
-
-export const taggingEndpointAdapter = endpointAdapterFactory(taggingEndpointAdapterOptions);
+});
