--- packages/x-adapter-platform/src/endpoint-adapters/top-clicked.endpoint-adapter.ts
@@ -1,16 +1,0 @@
-import { endpointAdapterFactory, EndpointAdapterOptions } from '@empathyco/x-adapter-next';
-import { BaseRequest } from '../types/request.types';
-import { TopClickedResponse } from '../types/response.types';
-import { topClickedResponseMapper } from '../mappers/response/top-clicked-response.mapper';
-import { baseRequestMapper } from '../mappers/request/base-request.mapper';
-
-export const topClickedEndpointAdapterOptions: EndpointAdapterOptions<
-  BaseRequest,
-  TopClickedResponse
-> = {
-  endpoint: 'https://api.{env(.)}empathy.co/search/v1/query/{instance}/topclicked',
-  responseMapper: topClickedResponseMapper,
-  requestMapper: baseRequestMapper
-};
-
-export const topClickedEndpointAdapter = endpointAdapterFactory(topClickedEndpointAdapterOptions);
