+++ packages/x-adapter-platform/src/mappers/response/__tests__/next-queries-response.mapper.spec.ts
@@ -1,8 +1,8 @@
 import { PlatformNextQueriesResponse } from '../../../types/responses/next-queries-response.model';
 import { nextQueriesResponseMapper } from '../next-queries-response.mapper';
 
-describe('next queries response mapper', () => {
-  it('should map the next queries', () => {
+describe('nextQueriesResponseMapper tests', () => {
+  it('should map the response', () => {
     const platformNextQueriesResponse: PlatformNextQueriesResponse = {
       data: {
         nextqueries: [
