+++ packages/x-adapter-platform/src/mappers/response/__tests__/search-response.mapper.spec.ts
@@ -1,9 +1,11 @@
 import { searchResponseMapper } from '../search-response.mapper';
 import { PlatformSearchResponse } from '../../../types/responses/search-response.model';
-import { platformSearchResponse } from '../../../__tests__/__fixtures__/platform-search.response';
-import { searchResponse } from '../../../__tests__/__fixtures__/search.response';
+import {
+  platformSearchResponse,
+  searchResponse
+} from '../../../__tests__/__fixtures__/platform-search.response';
 
-describe('search response mapper tests', () => {
+describe('searchResponseMapper tests', () => {
   it('should map the response', () => {
     expect(
       searchResponseMapper(platformSearchResponse as PlatformSearchResponse, {})
