--- packages/x-adapter-platform/src/mappers/response/empathize-response.mapper.ts
@@ -1,11 +1,0 @@
-import { schemaMapperFactory } from '@empathyco/x-adapter-next';
-import {
-  EmpathizeResponse,
-  PlatformEmpathizeResponse
-} from '../../types/responses/empathize-response.model';
-import { empathizeResponseMutableSchema } from '../../schemas/response/empathize-response.schema';
-
-export const empathizeResponseMapper = schemaMapperFactory<
-  PlatformEmpathizeResponse,
-  EmpathizeResponse
->(empathizeResponseMutableSchema);
