+++ packages/x-adapter-platform/src/mappers/response/identifier-results-response.mapper.ts
@@ -1,9 +1,9 @@
 import { schemaMapperFactory } from '@empathyco/x-adapter-next';
 import { IdentifierResultsResponse } from '@empathyco/x-types';
-/* eslint-disable max-len */
+// eslint-disable-next-line max-len
 import { PlatformIdentifierResultsResponse } from '../../types/responses/identifier-results-response.model';
+// eslint-disable-next-line max-len
 import { identifierResultsResponseMutableSchema } from '../../schemas/response/identifier-results-response.schema';
-/* eslint-enable max-len */
 
 export const identifierResultsResponseMapper = schemaMapperFactory<
   PlatformIdentifierResultsResponse,
