+++ packages/x-adapter-platform/src/mappers/response/index.ts
@@ -1,6 +1,7 @@
-export * from './empathize-response.mapper';
 export * from './identifier-results-response.mapper';
 export * from './next-queries-response.mapper';
+export * from './popular-searches-response.mapper';
+export * from './query-suggestions-response.mapper';
+export * from './recommendations-response.mapper';
 export * from './related-tags-response.mapper';
 export * from './search-response.mapper';
-export * from './top-clicked-response.mapper';
