+++ packages/x-adapter-platform/src/mappers/response/search-response.mapper.ts
@@ -1,8 +1,8 @@
 import { schemaMapperFactory } from '@empathyco/x-adapter-next';
 import { SearchResponse } from '@empathyco/x-types';
 import { PlatformSearchResponse } from '../../types/responses/search-response.model';
-import { mutableSearchResponseSchema } from '../../schemas/response/search-response.schema';
+import { searchResponseMutableSchema } from '../../schemas/response/search-response.schema';
 
 export const searchResponseMapper = schemaMapperFactory<PlatformSearchResponse, SearchResponse>(
-  mutableSearchResponseSchema
+  searchResponseMutableSchema
 );
