+++ packages/x-adapter-platform/src/schemas/index.ts
@@ -1,11 +1,2 @@
 export * from './request';
 export * from './response';
-export * from './result.schema';
-export * from './banner.schema';
-export * from './facet.schema';
-export * from './promoted.schema';
-export * from './redirection.schema';
-export * from './suggestion.schema';
-export * from './next-query.schema';
-export * from './related-tag.schema';
-export * from './filters';
