+++ packages/x-adapter-platform/src/schemas/response/identifier-results-response.schema.ts
@@ -1,7 +1,8 @@
 import { Schema, createMutableSchema } from '@empathyco/x-adapter-next';
 import { IdentifierResultsResponse } from '@empathyco/x-types';
-import { PlatformIdentifierResultsResponse } from '../../types';
-import { resultMutableSchema } from '../result.schema';
+// eslint-disable-next-line max-len
+import { PlatformIdentifierResultsResponse } from '../../types/responses/identifier-results-response.model';
+import { resultMutableSchema } from './models/result.schema';
 
 export const identifierResultsResponseSchema: Schema<
   PlatformIdentifierResultsResponse,
