+++ packages/x-adapter-platform/src/schemas/response/index.ts
@@ -1,6 +1,8 @@
-export * from './empathize-response.schema';
 export * from './identifier-results-response.schema';
+export * from './models';
 export * from './next-queries-response.schema';
+export * from './popular-searches-response.schema';
+export * from './query-suggestions-response.schema';
+export * from './recommendations-response.schema';
 export * from './related-tags-response.schema';
 export * from './search-response.schema';
-export * from './top-clicked-response.schema';
