+++ packages/x-adapter-platform/src/schemas/response/models/banner.schema.ts
@@ -1,7 +1,7 @@
 import { createMutableSchema, Schema } from '@empathyco/x-adapter-next';
 import { Banner } from '@empathyco/x-types';
 import { getTaggingInfoFromUrl } from '@empathyco/x-utils';
-import { PlatformBanner } from '../types/models.types';
+import { PlatformBanner } from '../../../types/responses/models/banner.model';
 
 export const bannerSchema: Schema<PlatformBanner, Banner> = {
   id: 'id',
