+++ packages/x-adapter-platform/src/schemas/response/models/index.ts
@@ -1,0 +1,8 @@
+export * from './banner.schema';
+export * from './facet.schema';
+export * from './next-query.schema';
+export * from './promoted.schema';
+export * from './redirection.schema';
+export * from './related-tag.schema';
+export * from './result.schema';
+export * from './suggestion.schema';
