+++ packages/x-adapter-platform/src/schemas/response/models/promoted.schema.ts
@@ -1,7 +1,7 @@
 import { Promoted } from '@empathyco/x-types';
 import { createMutableSchema, Schema } from '@empathyco/x-adapter-next';
 import { getTaggingInfoFromUrl } from '@empathyco/x-utils';
-import { PlatformPromoted } from '../types/models.types';
+import { PlatformPromoted } from '../../../types/responses/models/promoted.model';
 
 export const promotedSchema: Schema<PlatformPromoted, Promoted> = {
   id: 'id',
