+++ packages/x-adapter-platform/src/schemas/response/models/redirection.schema.ts
@@ -1,7 +1,7 @@
 import { createMutableSchema, Schema } from '@empathyco/x-adapter-next';
 import { Redirection } from '@empathyco/x-types';
 import { getTaggingInfoFromUrl } from '@empathyco/x-utils';
-import { PlatformRedirection } from '../types/models.types';
+import { PlatformRedirection } from '../../../types/responses/models/redirection.model';
 
 const redirectionSchema: Schema<PlatformRedirection, Redirection> = {
   id: 'id',
