+++ packages/x-adapter-platform/src/schemas/response/models/related-tag.schema.ts
@@ -1,6 +1,6 @@
 import { createMutableSchema, Schema } from '@empathyco/x-adapter-next';
 import { RelatedTag } from '@empathyco/x-types';
-import { PlatformRelatedTag } from '../types/models.types';
+import { PlatformRelatedTag } from '../../../types/responses/models/related-tag.model';
 
 export const relatedTagSchema: Schema<PlatformRelatedTag, RelatedTag> = {
   query: 'query',
